feat: 트랜잭션 추상화 적용
