+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -12,6 +12,7 @@ import com.techcourse.domain.User;
 import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
 import javax.sql.DataSource;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 
@@ -33,7 +34,7 @@ class UserServiceTest {
     @Test
     void testChangePassword() {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao);
+        final var userService = new AppUserService(userDao, userHistoryDao);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -45,10 +46,11 @@ class UserServiceTest {
     }
 
     @Test
+    @Disabled
     void testTransactionRollback() {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao);
+        final var userService = new AppUserService(userDao, userHistoryDao);
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
@@ -61,23 +63,23 @@ class UserServiceTest {
         assertThat(actual.getPassword()).isNotEqualTo(newPassword);
     }
 
-//    @Test
-//    void testTransactionRollback2() {
-//        // 트랜잭션 롤백 테스트를 위해 mock으로 교체
-//        final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-//        // 애플리케이션 서비스
-//        final var appUserService = new AppUserService(userDao, userHistoryDao);
-//        // 트랜잭션 서비스 추상화
-//        final var userService = new TxUserService(appUserService);
-//
-//        final var newPassword = "newPassword";
-//        final var createdBy = "gugu";
-//        // 트랜잭션이 정상 동작하는지 확인하기 위해 의도적으로 MockUserHistoryDao에서 예외를 발생시킨다.
-//        assertThrows(DataAccessException.class,
-//                () -> userService.changePassword(1L, newPassword, createdBy));
-//
-//        final var actual = userService.findById(1L);
-//
-//        assertThat(actual.getPassword()).isNotEqualTo(newPassword);
-//    }
+    @Test
+    void testTransactionRollback2() {
+        // 트랜잭션 롤백 테스트를 위해 mock으로 교체
+        final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
+        // 애플리케이션 서비스
+        final var appUserService = new AppUserService(userDao, userHistoryDao);
+        // 트랜잭션 서비스 추상화
+        final var userService = new TxUserService(appUserService);
+
+        final var newPassword = "newPassword";
+        final var createdBy = "gugu";
+        // 트랜잭션이 정상 동작하는지 확인하기 위해 의도적으로 MockUserHistoryDao에서 예외를 발생시킨다.
+        assertThrows(DataAccessException.class,
+                () -> userService.changePassword(1L, newPassword, createdBy));
+
+        final var actual = userService.findById(1L);
+
+        assertThat(actual.getPassword()).isNotEqualTo(newPassword);
+    }
 }
